+++ common/src/main/java/us/myles/ViaVersion/api/type/types/minecraft/PositionType.java
@@ -23,7 +23,7 @@ public class PositionType extends Type<Position> {
     @Override
     public void write(ByteBuf buffer, Position object) {
         buffer.writeLong((((long) object.getX() & 0x3ffffff) << 38)
-                | ((object.getY() & 0xfff) << 26)
+                | ((((long) object.getY()) & 0xfff) << 26)
                 | (object.getZ() & 0x3ffffff));
     }
 }
